% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character_data_frame.R
\name{as.character.data.frame}
\alias{as.character.data.frame}
\title{Convert all factors in a data.frame to characters}
\usage{
\method{as.character}{data.frame}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert all factors in a data.frame to characters
}

