% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_df.R
\name{character_df}
\alias{character_df}
\title{Convert all factors in a data.frame to characters}
\usage{
character_df(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert all factors in a data.frame to characters
}
