% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{connect to the unit test database}
\usage{
connect_db(host = "localhost", dbname = "n2kunittest",
  user = "unittest_analysis", password = "unittest", port = 5432)
}
\arguments{
\item{host}{the hostname of the database. Defaults to "localhost".}

\item{dbname}{the name of the unit test database. Defaults to "n2kunittest".}

\item{user}{the name of the unit test user. Defaults to "unittest_analysis".}

\item{password}{the password for the user. Defaults to "unittest".}

\item{port}{The port of host. Defaults to 5432.}
}
\description{
connect to the unit test database
}

