% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{Store objects of specific classes}
\usage{
store(object, conn, hash, clean = TRUE)
}
\arguments{
\item{object}{the object}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store objects of specific classes
}
