% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_analysis.R
\name{store_analysis}
\alias{store_analysis}
\title{store source species in the database}
\usage{
store_analysis(analysis, model_set, analysis_version, conn, hash,
  clean = TRUE)
}
\arguments{
\item{analysis}{as data.frame with file_fingerprint, model_set_local_id, location_group, species_group, last_year, seed, analysis_version, analysis_date, status and status_fingerprint.}

\item{model_set}{a data.frame with the model sets. Must have variables "local_id", "description", "first_year", "last_year" and "duration". The variable "long_description" is optional.}

\item{analysis_version}{an n2kAnalysisVersion object. See
\code{\link{get_analysis_version}}}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store source species in the database
}
