% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_analysis_dataset.R
\name{store_analysis_dataset}
\alias{store_analysis_dataset}
\title{store analysis and dataset in the database}
\usage{
store_analysis_dataset(analysis, model_set, analysis_version, dataset,
  analysis_dataset, clean = TRUE, hash, conn)
}
\arguments{
\item{analysis}{as data.frame with file_fingerprint, model_set_local_id, location_group, species_group, last_year, seed, analysis_version, analysis_date, status and status_fingerprint.}

\item{model_set}{a data.frame with the model sets. Must have variables "local_id", "description", "first_year", "last_year" and "duration". The variable "long_description" is optional.}

\item{analysis_version}{an n2kAnalysisVersion object. See
\code{\link{get_analysis_version}}}

\item{dataset}{a data.frame with names fingerprint, filename, datasource and import_date}

\item{analysis_dataset}{A \code{data.frame} linking the \code{file_fingerprint} from \code{analysis} to the \code{fingerprint} from \code{dataset}.}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}
}
\description{
store analysis and dataset in the database
}
