% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_analysis_version.R
\name{store_analysis_version}
\alias{store_analysis_version}
\title{Store the analysis version in the database}
\usage{
store_analysis_version(analysis_version, hash, clean = TRUE, conn)
}
\arguments{
\item{analysis_version}{an n2kAnalysisVersion object. See
\code{\link{get_analysis_version}}}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}

\item{conn}{a DBIconnection}
}
\description{
Store the analysis version in the database
}

