% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_anomaly_type.R
\name{store_anomaly_type}
\alias{store_anomaly_type}
\title{Store anomaly types}
\usage{
store_anomaly_type(anomaly_type, hash, conn, clean = TRUE)
}
\arguments{
\item{anomaly_type}{a data.frame with variables "local_id", "description" and "long_description". "long_description" is optional}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
Store anomaly types
}

