% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_connect_method.R
\name{store_connect_method}
\alias{store_connect_method}
\title{Store a vector of connect methods}
\usage{
store_connect_method(connect_method, hash, conn, clean = TRUE)
}
\arguments{
\item{connect_method}{the vector with connect methods.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
Store a vector of connect methods
}

