% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datafield.R
\name{store_datafield}
\alias{store_datafield}
\title{store a datafield in the database}
\usage{
store_datafield(datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{datafield}{a data.frame with datafield metadata. Must contain the variables local_id, datasource, table_name, primary_key and datafield_type. Other variables are ignored.}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store a datafield in the database
}

