% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datafield_type.R
\name{store_datafield_type}
\alias{store_datafield_type}
\title{Store a vector of datafield types}
\usage{
store_datafield_type(datafield_type, hash, conn, clean = TRUE)
}
\arguments{
\item{datafield_type}{the vector with datafield types.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store a vector of datafield types
}
