% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datasource.R
\name{store_datasource}
\alias{store_datasource}
\title{store a datasource in the database}
\usage{
store_datasource(datasource, conn)
}
\arguments{
\item{datasource}{a data.frame with datasource metadata}

\item{conn}{a DBIconnection}
}
\description{
store a datasource in the database
}
\details{
datasource must contain at least the variables description, datasource_type and connect_method.
}

