% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datasource_parameter.R
\name{store_datasource_parameter}
\alias{store_datasource_parameter}
\title{Store a vector of datasource parameters}
\usage{
store_datasource_parameter(datasource_parameter, hash, conn,
  clean = TRUE)
}
\arguments{
\item{datasource_parameter}{the vector with datasource parameters.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store a vector of datasource parameters
}
