% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_language.R
\name{store_language}
\alias{store_language}
\title{Store language}
\usage{
store_language(language, hash, conn, clean = TRUE)
}
\arguments{
\item{language}{the data.frame with language Must contains code and description. Other variables are ignored. code and description must have unique values.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
Store language
}

