% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_location.R
\name{store_location}
\alias{store_location}
\title{store locations in the database}
\usage{
store_location(location, datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{location}{a data.frame with location metadata. Must contain the following columns: local_id, description, parent_local_id, datafield_local_id and external_code. Other columns are ignored.}

\item{datafield}{a data.frame with datafield metadata. Must contain the variables local_id, datasource, table_name, primary_key and datafield_type. Other variables are ignored.}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store locations in the database
}
\details{
\itemize{
 \item location must have variables local_id, description, parent_local_id, datafield_local_id and extranal_code. Other variables are ignored
 \item datafield must have variables local_id, datasource, table_name, primary_key and datafield_type
 \item all local_id variables must be unique within their data.frame
 \item all values in location$datafield_local_id must exist in datafield$local_id
 \item all values in location$parent_location must be either NA or exist in location$local_id
}
}
