% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_location_group.R
\name{store_location_group}
\alias{store_location_group}
\title{Store location groups}
\usage{
store_location_group(location_group, hash, conn, clean = TRUE)
}
\arguments{
\item{location_group}{the data.frame with location groups. Must contains local_id, description and scheme. Other variables are ignored. local_id must have unique values.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
Store location groups
}

