% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_model_set.R
\name{store_model_set}
\alias{store_model_set}
\title{Store model sets in the database}
\usage{
store_model_set(model_set, hash, clean = TRUE, conn)
}
\arguments{
\item{model_set}{a data.frame with the model sets. Must have variables "local_id", "description", "first_year", "last_year" and "duration". The variable "long_description" is optional.}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}

\item{conn}{a DBIconnection}
}
\description{
Store model sets in the database
}

