% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_model_type.R
\name{store_model_type}
\alias{store_model_type}
\title{Store model type in the database}
\usage{
store_model_type(model_type, hash, clean = TRUE, conn)
}
\arguments{
\item{model_type}{a data.frame with the modeltypes. Must have a variable "description". The variable "long_description" is optional.}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}

\item{conn}{a DBIconnection}
}
\description{
Store model type in the database
}
