% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_n2kImport.R
\name{store_n2kImport}
\alias{store_n2kImport}
\title{store an n2kImport object into the database}
\usage{
store_n2kImport(object, conn, hash, clean = TRUE)
}
\arguments{
\item{object}{a \code{\link[n2kanalysis]{n2kImport-class}} object}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store an n2kImport object into the database
}
