% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_n2kManifest.R
\name{store_n2kManifest}
\alias{store_n2kManifest}
\title{store all models from an n2kManifest}
\usage{
store_n2kManifest(manifest, base, project, conn, status = "converged", hash,
  clean = TRUE)
}
\arguments{
\item{manifest}{a \code{\link[n2kanalysis]{n2kManifest-class}}}

\item{base}{the base location to read the model}

\item{project}{will be a relative path within the base location}

\item{conn}{a DBIconnection}

\item{status}{the status of the objects to be imported}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store all models from an n2kManifest
}
