% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_n2kResult.R
\name{store_n2kResult}
\alias{store_n2kResult}
\title{store an n2kResult object into the database}
\usage{
store_n2kResult(object, conn, hash, clean = TRUE)
}
\arguments{
\item{object}{a \code{\link[n2kanalysis]{n2kResult-class}} object}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store an n2kResult object into the database
}
