% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_observation.R
\name{store_observation}
\alias{store_observation}
\title{store a datafield in the database}
\usage{
store_observation(datafield, observation, location, parameter, conn, hash,
  clean = TRUE)
}
\arguments{
\item{datafield}{a data.frame with datafield metadata. Must contain the variables local_id, datasource, table_name, primary_key and datafield_type. Other variables are ignored.}

\item{observation}{a data.frame with observation metadata. Must contain the variables local_id, datafield_local_id, external_code, location_local_id, year and parameter_local_id. Other variables are ignored. datafield_local_id, external_code and parameter_local_id can be missing.}

\item{location}{a data.frame with location metadata. Must contain the following columns: local_id, description, parent_local_id, datafield_local_id and external_code. Other columns are ignored.}

\item{parameter}{a data.frame with parameters. Must contains the variables "description", "local_id", and "parent_parameter_local_id". Other variables are ignored.}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store a datafield in the database
}
