% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_parameter.R
\name{store_parameter}
\alias{store_parameter}
\title{Store parameters}
\usage{
store_parameter(parameter, hash, conn, clean = TRUE)
}
\arguments{
\item{parameter}{a data.frame with parameters. Must contains the variables "description", "local_id", and "parent_parameter_local_id". Other variables are ignored.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store parameters
}

