% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_scheme.R
\name{store_scheme}
\alias{store_scheme}
\title{Store a vector of schemes}
\usage{
store_scheme(scheme, hash, conn, clean = TRUE)
}
\arguments{
\item{scheme}{the vector with scheme descriptions.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store a vector of schemes
}
