% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_source_species.R
\name{store_source_species}
\alias{store_source_species}
\title{store source species in the database}
\usage{
store_source_species(source_species, datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{source_species}{a data.frame with source species metadata. Must contain local_id, description, datafield_local_id and extrenal_code. Other variables are ignored.}

\item{datafield}{a data.frame with datafield metadata. Must contain variables local_id, datasource, table_name, primary_key and datafield_type.}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
store source species in the database
}
