% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_source_species_species.R
\name{store_source_species_species}
\alias{store_source_species_species}
\title{store source species in the database}
\usage{
store_source_species_species(species, language, source_species,
  source_species_species, datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{species}{a data.frame with species metadata}

\item{language}{the data.frame with language Must contains code and description. Other variables are ignored. code and description must have unique values.}

\item{source_species}{a data.frame with source species metadata}

\item{source_species_species}{as data.frame linking the local species id to the local source_species id}

\item{datafield}{a data.frame with datafield metadata}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
store source species in the database
}
\details{
\itemize{
 \item species must contain at least the variables scientific_name and nbn_key. Other variables must be listed in language$code.
 \item source_species must have variables local_id, description, datafield_local_id and extranal_code. Other variables are ignored.
 \item datafield must have variables local_id, datasource, table_name, primary_key and datafield_type
 \item all local_id variables must be unique within their data.frame
 \item all values in source_species$datafield_local_id must exist in datafield$local_id
}
}

