% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_species.R
\name{store_species}
\alias{store_species}
\title{store species in the database}
\usage{
store_species(species, language, conn, hash, clean = TRUE)
}
\arguments{
\item{species}{a data.frame with species metadata. Must contain at least local_id, scientific_name and nbn_key. Other variable names must match the values in language$code..}

\item{language}{the data.frame with language Must contains code and description. Other variables are ignored. code and description must have unique values.}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
store species in the database
}

