% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_species_group_species.R
\name{store_species_group_species}
\alias{store_species_group_species}
\title{store all species related information in the database}
\usage{
store_species_group_species(species, language, source_species,
  source_species_species, datafield, species_group, species_group_species, conn)
}
\arguments{
\item{species}{a data.frame with species metadata}

\item{language}{the data.frame with language Must contains code and description. Other variables are ignored. code and description must have unique values.}

\item{source_species}{a data.frame with source species metadata}

\item{source_species_species}{as data.frame linking the local species id to the local source_species id}

\item{datafield}{a data.frame with datafield metadata}

\item{species_group}{the data.frame with species groups. Must contains local_id, description and scheme. Other variables are ignored. local_id must have unique values.}

\item{species_group_species}{as data.frame linking the local species group id to the local species id}

\item{conn}{a DBIconnection}
}
\description{
store all species related information in the database
}

