% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_public.R
\name{truncate_public}
\alias{truncate_public}
\title{Truncate all tables in the public schema: USE WITH CATION}
\usage{
truncate_public(conn)
}
\arguments{
\item{conn}{a DBIconnection}
}
\description{
Truncate all tables in the public schema: USE WITH CATION
}
