% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datafield.R
\name{store_datafield}
\alias{store_datafield}
\title{store a datafield in the database}
\usage{
store_datafield(datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{datafield}{a data.frame with datafield metadata}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
store a datafield in the database
}
\details{
datafield must contain the variables hash, datasource, table_name, primary_key and datafield_type. Other variables are ignored.
}

