% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datasource_type.R
\name{store_datasource_type}
\alias{store_datasource_type}
\title{Store a vector of datasource types}
\usage{
store_datasource_type(datasource_type, hash, conn, clean = TRUE)
}
\arguments{
\item{datasource_type}{the vector with connect methods.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
Store a vector of datasource types
}

