% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_datasource_type.R
\name{store_datasource_type}
\alias{store_datasource_type}
\title{Store a vector of datasource types}
\usage{
store_datasource_type(datasource_type, hash, conn, clean = TRUE)
}
\arguments{
\item{datasource_type}{the vector with datasource types.}

\item{hash}{the hash of the update session}

\item{conn}{a DBIconnection}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}
}
\description{
Store a vector of datasource types
}
