% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_location.R
\name{store_location}
\alias{store_location}
\title{store locations in the database}
\usage{
store_location(location, datafield, conn, hash, clean = FALSE)
}
\arguments{
\item{location}{a data.frame with location metadata}

\item{datafield}{a data.frame with datafield metadata}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
store locations in the database
}
\details{
\itemize{
 \item location mush have variables hash, description, parent_hash, datafield_hash and extranal_code. Other variables are ignored
 \item datafield mush have variables hash, datasource, table_name, primary_key and datafield_type
 \item all hash variables must be unique within their data.frame
 \item all values in location$datafield_hash must exist in datafield$hash
 \item all values in location$parent_location must be either NA or exist in location$hash
}
}

