% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_source_species.R
\name{store_source_species}
\alias{store_source_species}
\title{store source species in the database}
\usage{
store_source_species(source_species, datafield, conn, hash, clean = TRUE)
}
\arguments{
\item{source_species}{a data.frame with source species metadata}

\item{datafield}{a data.frame with datafield metadata}

\item{conn}{a DBIconnection}

\item{hash}{the hash of the update session}

\item{clean}{remove the staging table after update. Defaults to TRUE}
}
\description{
store source species in the database
}
\details{
\itemize{
 \item source_species must have variables local_id, description, datafield_local_id and extranal_code. Other variables are ignored.
 \item datafield must have variables local_id, datasource, table_name, primary_key and datafield_type
 \item all local_id variables must be unique within their data.frame
 \item all values in source_species$datafield_local_id must exist in datafield$local_id
}
}

