% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_species.R
\name{store_species}
\alias{store_species}
\title{store species in the database}
\usage{
store_species(species, language, conn)
}
\arguments{
\item{species}{a data.frame with species metadata}

\item{language}{the data.frame with language Must contains code and description. Other variables are ignored. code and description must have unique values.}

\item{conn}{a DBIconnection}
}
\description{
store species in the database
}
\details{
species must contain at least the variables scientific_name and nbn_key. Other variables must be listed in language$code.
}

