% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_status.R
\name{store_status}
\alias{store_status}
\title{Store status levels in the database}
\usage{
store_status(status, hash, clean = TRUE, conn)
}
\arguments{
\item{status}{a character vector with statuses}

\item{hash}{the hash of the update session}

\item{clean}{perform all database operations within a transaction and clean up the staging tables. Defaults to TRUE.}

\item{conn}{a DBIconnection}
}
\description{
Store status levels in the database
}

